feat(eror): add custom error type & reduce panics
This new error type makes the lib much safer, since it won't panic now
if there isn't a backlight. In theory the only remaining possible panics
are when the values in `brightness` and `max_brightness` cannot be
parsed, which should panic anyway if that occurs.

In tandem with these changes there have been some adjustments to main to
make it work with these errors and display them to the user without
panicking.